source folder -> sources folder
fixed a typo on folders name
