--- source/Main.cpp
@@ -1,16 +1,0 @@
-#include <iostream>
-#include "../headers/io.h"
-
-using namespace std;
-using namespace io;
-
-int main()
-{
-    Bienvenue();
-
-    cout << "Input de: ";
-	char r = de();
-	cout << "Result de : " << r << endl;
-
-	return 0;
-}
