+++ tests/io_main.cpp
@@ -1,10 +1,16 @@
 #include <iostream>
 #include "../headers/io.h"
+
 using namespace std;
 using namespace io;
+
 int main()
 {
+ 	Bienvenue();
+
+	cout << "Input de: ";
 	char r = de();
-	cout << "de : " << r << endl;
+	cout << "Result de : " << r << endl;
+
 	return 0;
 }
